kernel: Handle unknown VLAN in port PVID
The user may set the PVID to VID before adding a VLAN with VID to the
bridge.  Either for use as a poor-mans filtering of untagged traffic
on a trunk port, or to and the VLAN later.

When the VLAN is added, or later removed, we must update the port's
control buffer to point to either the (new) VLAN or NULL so the
forwarding stage can properly detect and drop incoming frames.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>